requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
matplotlib 
python-docx